Refactoring Component class World and Actor will be modify

+++ AnimProj/Map/GridMap.cpp
@@ -29,8 +29,8 @@ void pa::GridMap::clearMap()
 std::pair<size_t, size_t> pa::GridMap::getCellCoordinate(World& world, std::weak_ptr<Actor> actor) const
 {
 	const SceneComponent& sceneComp = actor.lock()->getComponent<SceneComponent>(world);
-	size_t cellX = sceneComp.position.x / _cellSize + _mapSize * 0.5;
-	size_t cellZ = sceneComp.position.z / _cellSize + _mapSize * 0.5;
+	size_t cellX = static_cast<size_t>(sceneComp.position.x / _cellSize + _mapSize * 0.5);
+	size_t cellZ = static_cast<size_t>(sceneComp.position.z / _cellSize + _mapSize * 0.5);
 
 	assert(cellX < _mapSize && cellZ < _mapSize);
 	return std::pair<size_t, size_t>(cellX, cellZ);
